worker: Improve log messages, e.g. avoid contractions

+++ lib/OpenQA/Worker/Engines/isotovideo.pm
@@ -142,24 +142,25 @@ sub cache_assets {
             log_debug($msg, channels => 'autoinst');
         }
 
+        die 'Mandatory $current_host is not set' unless $current_host;
         $asset = $cache_client->asset_path($current_host, $asset_uri)
           if $cache_client->asset_exists($current_host, $asset_uri);
 
         if ($this_asset eq 'UEFI_PFLASH_VARS' && !defined $asset) {
-            log_error("Can't download $asset_uri");
+            log_error("Failed to download $asset_uri");
             # assume that if we have a full path, that's what we should use
             $vars->{$this_asset} = $asset_uri if -e $asset_uri;
             # don't kill the job if the asset is not found
             # TODO: This seems to leave the job stuck in some cases (observed in production on openqaworker3).
             next;
         }
-        return {error => "Can't download $asset_uri to " . $cache_client->asset_path($current_host, $asset_uri)}
+        return {error => "Failed to download $asset_uri to " . $cache_client->asset_path($current_host, $asset_uri)}
           unless $asset;
         unlink basename($asset) if -l basename($asset);
         symlink($asset, basename($asset)) or die "cannot create link: $asset, $pooldir";
         $vars->{$this_asset} = path(getcwd, basename($asset))->to_string;
     }
-    return;
+    return undef;
 }
 
 sub engine_workit {
